+++ src/constants.ts
@@ -1,6 +1,9 @@
-import { SPAIN_COUNTRY } from "./geo/countries.js";
+import { COUNTRIES, GERMANY_COUNTRY, SPAIN_COUNTRY } from "./geo/countries.js";
 
 export const SPAIN_CENTER: [number, number] = SPAIN_COUNTRY.center;
+export const GERMANY_CENTER: [number, number] = GERMANY_COUNTRY.center;
+export const CZECHIA_CENTER: [number, number] = COUNTRIES.CZ.center;
+export const POLAND_CENTER: [number, number] = COUNTRIES.PL.center;
 export const DEFAULT_ZOOM = 6;
 
 export const DEFAULT_DRONE_PROFILE = {